M 优化发行默认支付渠道ID

+++ config/common.go
@@ -12,7 +12,7 @@ type CommonConfig struct {
 	TokenCryptKey           string   `mapstructure:"token_crypt_key" json:"token_crypt_key" yaml:"token_crypt_key"`                                  // jwt加密密钥
 	TokenSignKey            string   `mapstructure:"token_sign_key" json:"token_sign_key" yaml:"token_sign_key"`                                     // jwt内sign加密密钥
 	CommonHashKey           string   `mapstructure:"common_hash_key" json:"common_hash_key" yaml:"common_hash_key"`                                  // 常规hash加密密钥
-	PublishingPayChannelId  int      `mapstructure:"publishing_pay_channel_id" json:"publishing_pay_channel_id" yaml:"publishing_pay_channel_id"`    // 发行默认充值渠道ID
+	PublishingPayChannelId  int64    `mapstructure:"publishing_pay_channel_id" json:"publishing_pay_channel_id" yaml:"publishing_pay_channel_id"`    // 发行默认充值渠道ID
 	GameHashKey             string   `mapstructure:"game_hash_key" json:"game_hash_key" yaml:"game_hash_key"`                                        // 游戏加密密钥
 	CtxTokenDataKey         string   `mapstructure:"ctx_token_data_key" json:"ctx_token_data_key" yaml:"ctx_token_data_key"`                         // 上下文中存储token的键值
 	AuthorizationHeadKey    string   `mapstructure:"authorization_head_key" json:"authorization_head_key" yaml:"authorization_head_key"`             // 授权头部键值
